plant-doctor-service/requirements.txt
flask
boto3
numpy
pillow
python-dotenv
sagemaker>=2.218.0
